Updated at 22/08

+++ database/migrations/2016_08_11_171003_create_members_table.php
@@ -18,8 +18,8 @@ class CreateMembersTable extends Migration
             $table->foreign('family_id')->references('id')->on('families');
             $table->string('name', 100);
             $table->date('birthday')->nullable();
-            $table->integer('telephone')->nullable();
-            $table->integer('cellphone')->nullable();
+            $table->bigInteger('telephone')->nullable();
+            $table->bigInteger('cellphone')->nullable();
             $table->string('email', 60)->nullable();
             $table->char('gender',1)->nullable();
             $table->boolean('status')->default('1');
